i37 - JSON input - split the interfaces and JSON impl
