+++ src/main/kotlin/cz/dynawest/csvcruncher/converters/TabularPropertiesMetadataCollector.kt
@@ -1,0 +1,8 @@
+package cz.dynawest.csvcruncher.converters
+
+class TabularPropertiesMetadataCollector : EntryProcessor {
+    val propertiesSoFar: Map<String, PropertyInfo> = hashMapOf()
+    override fun process(entry: Entry) {
+        TODO("Not yet implemented")
+    }
+}
